+++ shared/editor/nodes/BulletList.ts
@@ -1,40 +1,13 @@
-import { wrappingInputRule, InputRule } from "prosemirror-inputrules";
 import {
   Schema,
   NodeType,
   NodeSpec,
   Node as ProsemirrorModel,
-  Attrs,
 } from "prosemirror-model";
 import toggleList from "../commands/toggleList";
 import { MarkdownSerializerState } from "../lib/markdown/serializer";
-import { isInHeading } from "../queries/isInHeading";
+import { listWrappingInputRule } from "../lib/listInputRule";
 import Node from "./Node";
-
-/**
- * A wrapper for wrappingInputRule that prevents execution inside heading nodes.
- * This fixes the bug where typing "* " or "- " inside a heading would trigger bullet list conversion.
- */
-function safeWrappingInputRule(
-  regexp: RegExp,
-  nodeType: NodeType,
-  getAttrs?: (match: RegExpMatchArray) => Attrs | null,
-  joinPredicate?: (match: RegExpMatchArray, node: ProsemirrorModel) => boolean
-): InputRule {
-  const rule = wrappingInputRule(regexp, nodeType, getAttrs, joinPredicate);
-
-  // Wrap the original rule to check if we're inside a heading
-  return new InputRule(regexp, (state, match, start, end) => {
-    // Don't apply the rule if we're inside a heading
-    if (isInHeading(state)) {
-      return null;
-    }
-
-    // Otherwise, execute the original wrappingInputRule handler
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    return (rule as any).handler(state, match, start, end);
-  });
-}
 
 export default class BulletList extends Node {
   get name() {
@@ -61,7 +34,7 @@ export default class BulletList extends Node {
   }
 
   inputRules({ type }: { type: NodeType }) {
-    return [safeWrappingInputRule(/^\s*([-+*])\s$/, type)];
+    return [listWrappingInputRule(/^\s*([-+*])\s$/, type)];
   }
 
   toMarkdown(state: MarkdownSerializerState, node: ProsemirrorModel) {
