+++ shared/editor/nodes/OrderedList.ts
@@ -1,41 +1,14 @@
 import { Token } from "markdown-it";
-import { wrappingInputRule, InputRule } from "prosemirror-inputrules";
 import {
   NodeSpec,
   NodeType,
   Schema,
   Node as ProsemirrorNode,
-  Attrs,
 } from "prosemirror-model";
 import toggleList from "../commands/toggleList";
 import { MarkdownSerializerState } from "../lib/markdown/serializer";
-import { isInHeading } from "../queries/isInHeading";
+import { listWrappingInputRule } from "../lib/listInputRule";
 import Node from "./Node";
-
-/**
- * A wrapper for wrappingInputRule that prevents execution inside heading nodes.
- * This fixes the bug where typing "1. " inside a heading would trigger ordered list conversion.
- */
-function safeWrappingInputRule(
-  regexp: RegExp,
-  nodeType: NodeType,
-  getAttrs?: (match: RegExpMatchArray) => Attrs | null,
-  joinPredicate?: (match: RegExpMatchArray, node: ProsemirrorNode) => boolean
-): InputRule {
-  const rule = wrappingInputRule(regexp, nodeType, getAttrs, joinPredicate);
-
-  // Wrap the original rule to check if we're inside a heading
-  return new InputRule(regexp, (state, match, start, end) => {
-    // Don't apply the rule if we're inside a heading
-    if (isInHeading(state)) {
-      return null;
-    }
-
-    // Otherwise, execute the original wrappingInputRule handler
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    return (rule as any).handler(state, match, start, end);
-  });
-}
 
 export default class OrderedList extends Node {
   get name() {
@@ -106,14 +79,14 @@ export default class OrderedList extends Node {
 
   inputRules({ type }: { type: NodeType }) {
     return [
-      safeWrappingInputRule(
+      listWrappingInputRule(
         /^(\d+)\.\s$/,
         type,
         (match) => ({ order: +match[1], listStyle: "number" }),
         (match, node) => node.childCount + node.attrs.order === +match[1]
       ),
 
-      safeWrappingInputRule(
+      listWrappingInputRule(
         /^([a-z])\.\s$/,
         type,
         (match) => ({
@@ -128,7 +101,7 @@ export default class OrderedList extends Node {
         }
       ),
 
-      safeWrappingInputRule(
+      listWrappingInputRule(
         /^([A-Z])\.\s$/,
         type,
         (match) => ({
